Restructure source code and modified readme accordingly
