+++ DirectProgramming/C++SYCL_FPGA/Tutorials/Tools/use_library/part2_lib_func_call/src/lib_rtl_spec.xml
@@ -1,9 +1,9 @@
 <RTL_SPEC>
-  <FUNCTION name="RtlByteswap" module="byteswap_uint">
+  <FUNCTION name="RtlDSPm27x27u" module="dsp_27x27u">
     <ATTRIBUTES>
       <IS_STALL_FREE value="yes"/>
       <IS_FIXED_LATENCY value="yes"/>
-      <EXPECTED_LATENCY value="0"/>
+      <EXPECTED_LATENCY value="2"/>
       <CAPACITY value="0"/>
       <HAS_SIDE_EFFECTS value="no"/>
       <ALLOW_MERGING value="yes"/>
@@ -15,11 +15,15 @@
       <AVALON port="iready" type="iready"/>
       <AVALON port="ovalid" type="ovalid"/>
       <AVALON port="oready" type="oready"/>
-      <INPUT  port="datain" width="32"/>
-      <OUTPUT port="dataout" width="32"/>
+      <INPUT  port="ax" width="27"/>
+      <INPUT  port="ay" width="27"/>
+      <OUTPUT port="resulta" width="54"/>
     </INTERFACE>
     <REQUIREMENTS>
       <FILE name="lib_rtl.v" />
     </REQUIREMENTS>
+    <RESOURCES>
+      <DSPS value="1"/>
+    </RESOURCES>
   </FUNCTION>
 </RTL_SPEC>
