--- DirectProgramming/C++SYCL_FPGA/Tutorials/Tools/use_library/src/lib.hpp
@@ -1,9 +1,0 @@
-// =============================================================
-// Copyright Intel Corporation
-//
-// SPDX-License-Identifier: MIT
-// =============================================================
-#include <sycl/sycl.hpp>
-
-SYCL_EXTERNAL extern "C" unsigned RtlByteswap(unsigned x);
-
