--- DirectProgramming/C++SYCL_FPGA/Tutorials/Tools/use_library/src/lib_rtl_model.cpp
@@ -1,11 +1,0 @@
-//==============================================================
-// Copyright Intel Corporation
-//
-// SPDX-License-Identifier: MIT
-// =============================================================
-#include <sycl/sycl.hpp>
-
-SYCL_EXTERNAL extern "C" unsigned RtlByteswap (unsigned x) {
-  return x << 16 | x >> 16;
-}
-
